greengrass-observability/requirements.txt
awsiotsdk==1.11.8